+++ libs/community/langchain_community/tools/passio_nutrition_ai/__init__.py
@@ -1,0 +1,5 @@
+"""Passio Nutrition AI API toolkit."""
+
+from langchain_community.tools.passio_nutrition_ai.tool import NutritionAI
+
+__all__ = ["NutritionAI"]
