--- libs/community/langchain_community/tools/passio_nutritionai/__init__.py
@@ -1,5 +1,0 @@
-"""Passio Nutrition AI API toolkit."""
-
-from langchain_community.tools.passio_nutritionai.tool import NutritionAI
-
-__all__ = ["NutritionAI"]
